backend/requirements.txt
# SpeechRecognition
# pydub
# ffmpeg-python
# moviepy
openai-whisper
pandas
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
google-auth
PyPDF2 
python-docx
pydantic
openai
aiofiles
fastapi
uvicorn[standard]
python-multipart
supabase==2.9.1
gotrue==2.9.2
python-dotenv
# SQLAlchemy dependencies - keeping for compatibility but not using
sqlalchemy
alembic
asyncpg
cryptography
sqlalchemy[asyncio]